feat: extract executeMutation() helper, add custom macros docs
Extract the duplicated clone-execute-diff-apply pattern from Set, Do,
Button, and MacroLink into a shared executeMutation() function. Add
authoring documentation for creating custom macros with the registry,
covering state reading, mutation, and execution timing. Update widget
docs to document [widget]-tagged passages.

Closes #19

release-npm
