+++ src/components/macros/Do.tsx
@@ -1,10 +1,8 @@
 import { useLayoutEffect, useContext } from 'preact/hooks';
-import { useStoryStore } from '../../store';
-import { execute } from '../../expression';
 import type { ASTNode } from '../../markup/ast';
-import { deepClone } from '../../class-registry';
 import { LocalsContext } from '../../markup/render';
 import { useMergedLocals } from '../../hooks/use-merged-locals';
+import { executeMutation } from '../../execute-mutation';
 import { currentSourceLocation } from '../../utils/source-location';
 
 interface DoProps {
@@ -24,33 +22,10 @@ export function Do({ children }: DoProps) {
   const [, , mergedLocals] = useMergedLocals();
 
   useLayoutEffect(() => {
-    const state = useStoryStore.getState();
-    const vars = deepClone(state.variables);
-    const temps = deepClone(state.temporary);
-    const localsClone = { ...mergedLocals };
-
     try {
-      execute(code, vars, temps, localsClone);
+      executeMutation(code, mergedLocals, scope.update);
     } catch (err) {
       console.error(`spindle: Error in {do}${currentSourceLocation()}:`, err);
-      return;
-    }
-
-    // Diff and apply
-    for (const key of Object.keys(vars)) {
-      if (vars[key] !== state.variables[key]) {
-        state.setVariable(key, vars[key]);
-      }
-    }
-    for (const key of Object.keys(temps)) {
-      if (temps[key] !== state.temporary[key]) {
-        state.setTemporary(key, temps[key]);
-      }
-    }
-    for (const key of Object.keys(localsClone)) {
-      if (localsClone[key] !== mergedLocals[key]) {
-        scope.update(`@${key}`, localsClone[key]);
-      }
     }
   }, []);
