+++ src/components/macros/Set.tsx
@@ -1,9 +1,7 @@
 import { useRef, useContext } from 'preact/hooks';
-import { useStoryStore } from '../../store';
-import { execute } from '../../expression';
-import { deepClone } from '../../class-registry';
 import { LocalsContext } from '../../markup/render';
 import { useMergedLocals } from '../../hooks/use-merged-locals';
+import { executeMutation } from '../../execute-mutation';
 import { currentSourceLocation } from '../../utils/source-location';
 
 interface SetProps {
@@ -17,38 +15,15 @@ export function Set({ rawArgs }: SetProps) {
 
   if (!ran.current) {
     ran.current = true;
-    const state = useStoryStore.getState();
-    const vars = deepClone(state.variables);
-    const temps = deepClone(state.temporary);
-    const localsClone = { ...mergedLocals };
 
     try {
-      execute(rawArgs, vars, temps, localsClone);
+      executeMutation(rawArgs, mergedLocals, scope.update);
     } catch (err) {
       console.error(
         `spindle: Error in {set ${rawArgs}}${currentSourceLocation()}:`,
         err,
       );
       return null;
-    }
-
-    // Diff and apply store changes
-    for (const key of Object.keys(vars)) {
-      if (vars[key] !== state.variables[key]) {
-        state.setVariable(key, vars[key]);
-      }
-    }
-    for (const key of Object.keys(temps)) {
-      if (temps[key] !== state.temporary[key]) {
-        state.setTemporary(key, temps[key]);
-      }
-    }
-
-    // Diff and apply locals changes
-    for (const key of Object.keys(localsClone)) {
-      if (localsClone[key] !== mergedLocals[key]) {
-        scope.update(`@${key}`, localsClone[key]);
-      }
     }
   }
