+++ src/execute-mutation.ts
@@ -1,0 +1,32 @@
+import { useStoryStore } from './store';
+import { execute } from './expression';
+import { deepClone } from './class-registry';
+
+export function executeMutation(
+  code: string,
+  mergedLocals: Record<string, unknown>,
+  scopeUpdate: (key: string, value: unknown) => void,
+): void {
+  const state = useStoryStore.getState();
+  const vars = deepClone(state.variables);
+  const temps = deepClone(state.temporary);
+  const localsClone = { ...mergedLocals };
+
+  execute(code, vars, temps, localsClone);
+
+  for (const key of Object.keys(vars)) {
+    if (vars[key] !== state.variables[key]) {
+      state.setVariable(key, vars[key]);
+    }
+  }
+  for (const key of Object.keys(temps)) {
+    if (temps[key] !== state.temporary[key]) {
+      state.setTemporary(key, temps[key]);
+    }
+  }
+  for (const key of Object.keys(localsClone)) {
+    if (localsClone[key] !== mergedLocals[key]) {
+      scopeUpdate(`@${key}`, localsClone[key]);
+    }
+  }
+}
